Move the rest of my components to my client directory.
