+++ src/clients/avalonstar/components/index.js
@@ -1,5 +1,6 @@
 export { default as ActivityCamera } from './Cameras';
 export { default as EmoteCounter } from './Emotes';
 export { Panel, Notifier, Ticker } from './Events';
+export { default as Generic } from './Generic';
 export { default as SubPoints } from './SubPoints';
 export { default as Uptime } from './Uptime';
