--- src/components/Labels/index.js
@@ -1,1 +1,0 @@
-export { default as Generic } from './Generic';
